+++ pages/mypage/edit/index.tsx
@@ -2,7 +2,7 @@ import MyPageEditPage from "../../../components/pages/Mypage/Edit";
 
 export default function MyPageEdit() {
   const data = {
-    avatarSrc: "/img/ProjectDefaultBackground.png",
+    avatarSrc: "/images/ProjectDefaultBackground.png",
     nickname: "자라는개발자",
     introduction:
       "프론트엔드 개발자를 꿈꾸는 취준생입니다. 프로젝트 경험하고 싶어요",
@@ -11,7 +11,7 @@ export default function MyPageEdit() {
     following: 30,
     skill: ["typescript", "react", "HTML", "Next.js", "React.native"],
     position: "프론트엔드",
-    career: "신입",
+    career: "0",
     github: "https://github.com",
     blog: "https://www.naver.com",
     portfolio: "https://www.naver.com",
